Split out common static functions to libsocketcan-utils.c
libsocketcan.c has some common function between libcangw.c.  But
that common function is implemented with static.

This patch split out common static function to libsocketcan-utils.c.
Each function remove static from function.  In addition, it's
added __attribute__((__visibility__("hidden"))) to avoid symbol export.
