+++ src/GNUmakefile.am
@@ -4,7 +4,9 @@ AM_CPPFLAGS = \
 	-I$(top_srcdir)/include \
 	-I$(top_builddir)/include
 
-libsocketcan_la_SOURCES = libsocketcan.c
+libsocketcan_la_SOURCES = \
+	libsocketcan-utils.c \
+	libsocketcan.c
 
 libsocketcan_la_LDFLAGS = \
 	-version-info $(LT_CURRENT):$(LT_REVISION):$(LT_AGE)
